move shaders page to main threejs journey page
