--- pages/notes/threejs-journey/shaders/index.tsx
@@ -1,155 +1,0 @@
-import Head from "next/head";
-import { Sandpack } from "@codesandbox/sandpack-react";
-import { atomDark } from "@codesandbox/sandpack-themes";
-import {
-  ArticleContainer,
-  Lede,
-  Navbar,
-  Paragraph,
-  Subheader,
-  Title,
-} from "../../../../components";
-import ShaderPatternsCodeSandbox from "../../../../interactives/notes/threejs-journey/29-shader-patterns";
-import { shaderPatterns } from "../../../../interactives/notes/threejs-journey/29-shader-patterns/shader-patterns";
-import {
-  indexHtmlFile,
-  indexJsFile,
-  stylesCssFile,
-  vertexGlslFile,
-  fragmentGlslFile,
-} from "../../../../interactives/notes/threejs-journey/28-shaders";
-import { useState } from "react";
-import * as Select from "@radix-ui/react-select";
-import { FiCheck, FiChevronDown } from "react-icons/fi";
-import _ from "lodash";
-
-const ShadersPage = () => {
-  const [selectedShaderPattern, setSelectedShaderPattern] = useState(
-    shaderPatterns[0]
-  );
-  const onValueChange = (patternName) => {
-    const patternObj = _.find(shaderPatterns, (pattern) => {
-      return patternName === pattern.name;
-    });
-
-    setSelectedShaderPattern(patternObj);
-  };
-
-  const files = {
-    "/index.html": {
-      code: indexHtmlFile(),
-      hidden: true,
-    },
-    "/index.js": indexJsFile(`${vertexGlslFile()}`, fragmentGlslFile()),
-    "/styles.css": {
-      code: stylesCssFile(),
-      hidden: true,
-    },
-    "/vertex.glsl": vertexGlslFile(),
-    "/fragment.glsl": fragmentGlslFile(),
-  };
-
-  return (
-    <>
-      <Head>
-        <title>ThreeJS Journey Notes / Newt Interactive</title>
-        <meta
-          name="description"
-          content="Interactive, educational explainers and playgrounds on topics in science, technology, engineering, and math"
-        />
-        {/* <!-- Twitter --> */}
-        <meta property="twitter:card" content="summary_large_image" />
-        <meta
-          property="twitter:url"
-          content="https://www.newtinteractive.com/notes/threejs-journey"
-        />
-        <meta property="twitter:creator" content="@nehaludyavar" />
-        <meta
-          property="twitter:title"
-          content="ThreeJS Journey Notes / Newt Interactive"
-        />
-        <meta
-          property="twitter:description"
-          content="Interactive, educational explainers and playgrounds on topics in science, technology, engineering, and math"
-        />
-        <meta
-          property="twitter:image"
-          content="https://i.ibb.co/NpXN5rj/Meta-tag-image-1.png"
-        />
-      </Head>
-      <Navbar />
-      <ArticleContainer>
-        <Title>ThreeJS Journey Notes</Title>
-        <Lede>
-          Random stuff from Bruno Simon's course,{" "}
-          <a
-            href="https://threejs-journey.com/"
-            target="_blank"
-            rel="noreferrer noopener"
-            className="text-slate-600 hover:text-slate-700 underline underline-offset-1 decoration-slate-700"
-          >
-            ThreeJS Journey
-          </a>
-        </Lede>
-        <Subheader>28 &mdash; Shaders</Subheader>
-        <Paragraph>Shaders</Paragraph>
-        <Sandpack
-          files={files}
-          theme={atomDark}
-          template="vanilla"
-          options={{
-            showLineNumbers: true,
-            editorHeight: 500,
-          }}
-          customSetup={{
-            dependencies: {
-              three: "0.151.3",
-            },
-            devDependencies: {
-              "@babel/core": "7.2.0",
-              "parcel-bundler": "^1.6.1",
-            },
-            entry: "index.html",
-          }}
-        />
-        <Subheader className="mt-20">29 &mdash; Shader patterns</Subheader>
-        <Select.Root
-          value={selectedShaderPattern.name}
-          onValueChange={onValueChange}
-        >
-          <Select.Trigger className="inline-flex leading-none rounded data-[placeholder]:text-slate-600 outline-none text-sm py-1">
-            <Select.Value aria-label={selectedShaderPattern.name}>
-              {selectedShaderPattern.name}
-            </Select.Value>
-            <Select.Icon>
-              <FiChevronDown className="ml-1" />
-            </Select.Icon>
-          </Select.Trigger>
-          <Select.Portal>
-            <Select.Content
-              position="popper"
-              className="overflow-hidden bg-white rounded-md shadow-[0px_10px_38px_-10px_rgba(22,_23,_24,_0.35),0px_10px_20px_-15px_rgba(22,_23,_24,_0.2)]"
-            >
-              <Select.Viewport className="p-[5px]">
-                {shaderPatterns.map((pattern) => (
-                  <Select.Item
-                    className="text-sm leading-none rounded-sm flex items-center h-6 pr-9 pl-6 relative select-none data-[disabled]:text-slate-400 data-[disabled]:pointer-events-none data-[highlighted]:outline-none data-[highlighted]:bg-slate-600 data-[highlighted]:text-white"
-                    value={pattern.name}
-                  >
-                    <Select.ItemText>{pattern.name}</Select.ItemText>
-                    <Select.ItemIndicator className="absolute left-0 w-[25px] inline-flex items-center justify-center">
-                      <FiCheck />
-                    </Select.ItemIndicator>
-                  </Select.Item>
-                ))}
-              </Select.Viewport>
-            </Select.Content>
-          </Select.Portal>
-        </Select.Root>
-        <ShaderPatternsCodeSandbox shaderPattern={selectedShaderPattern} />
-      </ArticleContainer>
-    </>
-  );
-};
-
-export default ShadersPage;
